requirements.txt
boto3==1.2.1
click==5.1
pyhorn==0.5.0
pyloggly==0.1.0
python-dotenv==0.3.0
stopit==1.1.1
Unipath==1.1
requests==2.8.1
nose==1.3.7
